[SIMULIZAR-115] Incorporate review
